hotfix/revisao | ajustes requisitados pelo grupo 1
- Funcionaliade da sidebar asjutada para o scrool ser somente na lista
  de pedidos
- Deleção de comentarios e codigos desnecessários
- Deleção da seleção de cartão
	- O codigo estava conflitante ao ponto de não ser utilizavel por
	  esse motivo optamos por deleta-lo
