+++ lib/app/modules/cart/cart_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:pscomidas/app/modules/cart/cart_store.dart';
 import 'package:flutter/material.dart';
-import 'package:pscomidas/app/modules/cart/components/cart_drawer/cart_drawer.dart';
+import 'package:pscomidas/app/modules/cart/components/drawer/cart_drawer.dart';
 
 class CartPage extends StatefulWidget {
   final String title;
